+++ experiments/tester/src/RNTester/TestComponents/index.ts
@@ -7,6 +7,7 @@ import { TextTest } from './Text/TextTest';
 import { ThemeTest } from './ThemeTest';
 import { FocusTrapTest } from './FocusTrapZoneTest';
 import { PersonaCoinTest } from './PersonaCoin/PersonaCoinTest';
+import { RadioGroupTest } from './RadioGroupTest';
 import { PersonaTest } from './Persona/PersonaTest';
 
 export type TestDescription = {
@@ -46,6 +47,10 @@ export const allTestComponents: TestDescription[] = [
   {
     name: 'PersonaCoin Test',
     component: PersonaCoinTest
+  },
+  {
+    name: 'RadioGroup Test',
+    component: RadioGroupTest
   },
   {
     name: 'Persona Test',