+++ packages/utils/interactive-hooks/src/index.ts
@@ -4,3 +4,4 @@ export * from './useFocusState';
 export * from './useHoverState';
 export * from './usePressState';
 export * from './useViewCommandFocus';
+export * from './RadioGroup.hooks';