+++ scripts/configs/index.js
@@ -1,3 +1,4 @@
 module.exports = {
+  ...require('./configure-jest'),
   ...require('./configure-metro')
 };